+++ app/src/main/java/eu/kanade/presentation/more/settings/screen/SettingsTrackingScreen.kt
@@ -8,6 +8,9 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.text.input.TextFieldLineLimits
+import androidx.compose.foundation.text.input.TextObfuscationMode
+import androidx.compose.foundation.text.input.rememberTextFieldState
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
@@ -16,6 +19,7 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.OutlinedSecureTextField
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -38,9 +42,6 @@ import androidx.compose.ui.semantics.contentType
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import dev.icerock.moko.resources.StringResource
@@ -311,8 +312,8 @@ object SettingsTrackingScreen : SearchableSettings {
         val context = LocalContext.current
         val scope = rememberCoroutineScope()
 
-        var username by remember { mutableStateOf(TextFieldValue(tracker.getUsername())) }
-        var password by remember { mutableStateOf(TextFieldValue(tracker.getPassword())) }
+        val username = rememberTextFieldState(tracker.getUsername())
+        val password = rememberTextFieldState(tracker.getPassword())
         var processing by remember { mutableStateOf(false) }
         var inputError by remember { mutableStateOf(false) }
 
@@ -338,21 +339,19 @@ object SettingsTrackingScreen : SearchableSettings {
                         modifier = Modifier
                             .fillMaxWidth()
                             .semantics { contentType = ContentType.Username + ContentType.EmailAddress },
-                        value = username,
-                        onValueChange = { username = it },
+                        state = username,
                         label = { Text(text = stringResource(uNameStringRes)) },
                         keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
-                        singleLine = true,
+                        lineLimits = TextFieldLineLimits.SingleLine,
                         isError = inputError && !processing,
                     )
 
                     var hidePassword by remember { mutableStateOf(true) }
-                    OutlinedTextField(
+                    OutlinedSecureTextField(
                         modifier = Modifier
                             .fillMaxWidth()
                             .semantics { contentType = ContentType.Password },
-                        value = password,
-                        onValueChange = { password = it },
+                        state = password,
                         label = { Text(text = stringResource(MR.strings.password)) },
                         trailingIcon = {
                             IconButton(onClick = { hidePassword = !hidePassword }) {
@@ -366,16 +365,15 @@ object SettingsTrackingScreen : SearchableSettings {
                                 )
                             }
                         },
-                        visualTransformation = if (hidePassword) {
-                            PasswordVisualTransformation()
+                        textObfuscationMode = if (hidePassword) {
+                            TextObfuscationMode.Hidden
                         } else {
-                            VisualTransformation.None
+                            TextObfuscationMode.Visible
                         },
                         keyboardOptions = KeyboardOptions(
                             keyboardType = KeyboardType.Password,
                             imeAction = ImeAction.Done,
                         ),
-                        singleLine = true,
                         isError = inputError && !processing,
                     )
                 }
@@ -390,8 +388,8 @@ object SettingsTrackingScreen : SearchableSettings {
                             val result = checkLogin(
                                 context = context,
                                 tracker = tracker,
-                                username = username.text,
-                                password = password.text,
+                                username = username.text.toString(),
+                                password = password.text.toString(),
                             )
                             inputError = !result
                             if (result) onDismissRequest()
@@ -503,7 +501,7 @@ object SettingsTrackingScreen : SearchableSettings {
         val uriHandler = LocalUriHandler.current
         val scope = rememberCoroutineScope()
 
-        var token by remember { mutableStateOf(TextFieldValue(tracker.getPassword())) }
+        val token = rememberTextFieldState(tracker.getPassword())
         var processing by remember { mutableStateOf(false) }
         var inputError by remember { mutableStateOf(false) }
 
@@ -546,12 +544,11 @@ object SettingsTrackingScreen : SearchableSettings {
                     )
 
                     var hideToken by remember { mutableStateOf(true) }
-                    OutlinedTextField(
+                    OutlinedSecureTextField(
                         modifier = Modifier
                             .fillMaxWidth()
                             .semantics { contentType = ContentType.Password },
-                        value = token,
-                        onValueChange = { token = it },
+                        state = token,
                         label = { Text(text = stringResource(tokenStringRes)) },
                         trailingIcon = {
                             IconButton(onClick = { hideToken = !hideToken }) {
@@ -565,16 +562,15 @@ object SettingsTrackingScreen : SearchableSettings {
                                 )
                             }
                         },
-                        visualTransformation = if (hideToken) {
-                            PasswordVisualTransformation()
+                        textObfuscationMode = if (hideToken) {
+                            TextObfuscationMode.Hidden
                         } else {
-                            VisualTransformation.None
+                            TextObfuscationMode.Visible
                         },
                         keyboardOptions = KeyboardOptions(
                             keyboardType = KeyboardType.Password,
                             imeAction = ImeAction.Done,
                         ),
-                        singleLine = true,
                         isError = inputError && !processing,
                     )
 
@@ -596,7 +592,7 @@ object SettingsTrackingScreen : SearchableSettings {
                                 context = context,
                                 tracker = tracker,
                                 username = "",
-                                password = token.text,
+                                password = token.text.toString(),
                             )
                             inputError = !result
                             if (result) onDismissRequest()